third_party/blink/web_tests/fast/mediarecorder/MediaRecorder-disabled-tracks.html

<!DOCTYPE html>
<script src=../../resources/testharness.js></script>
<script src=../../resources/testharnessreport.js></script>
<script>

// This test verifies that MediaStream with disabled tracks can be correctly
// recorded. See crbug.com/878255 for more context.
var makeAsyncTest = function(value, expected) {
  var recorder;

  async_test(function(test) {
    const recorderOnDataAvailable = this.step_func(function(event) {
      if (recorder.state != "recording")
        return;

      recorder.onstop = recorderOnStopExpected;
      recorder.stop();
    });

    const recorderOnStopExpected = this.step_func_done();
    const recorderOnStopUnexpected = test.unreached_func('Recording stopped.');
    const recorderOnError = test.unreached_func('Recording error.');

    const gotStream = this.step_func(function(stream) {
      for (track in stream.getTracks())
        track.enabled = true;

      recorder = new MediaRecorder(stream);

      assert_equals(recorder.state, "inactive");
      recorder.ondataavailable = recorderOnDataAvailable;
      recorder.onstop = recorderOnStopUnexpected;
      recorder.onerror = recorderOnError;
      recorder.start();

      assert_equals(recorder.state, "recording");
      recorder.requestData();
    });

    const onError = test.unreached_func('Error creating MediaStream.');
    navigator.webkitGetUserMedia(value, gotStream, onError);
  });
};

generate_tests(makeAsyncTest,
               [["video-only",  {video: true,  audio: false}],
                ["audio-only",  {video: false, audio: true}],
                ["audio-video", {video: true,  audio: true}]]);

</script>
